+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.21.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.20.1...@synapsecns/synapse-interface@0.21.0) (2024-04-30)
+
+
+### Features
+
+* **synapse-interface:** maintenance aggregator using PAUSED_CHAINS ([#2345](https://github.com/synapsecns/sanguine/issues/2345)) ([ee14cb8](https://github.com/synapsecns/sanguine/commit/ee14cb82e04ec16d6fccc8cfb7f3194a1b99c140)), closes [#2346](https://github.com/synapsecns/sanguine/issues/2346)
+
+
+
+
+
+## [0.20.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.20.0...@synapsecns/synapse-interface@0.20.1) (2024-04-29)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+# [0.20.0](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.19.1...@synapsecns/synapse-interface@0.20.0) (2024-04-29)
+
+
+### Features
+
+* **synapse-interface:** unpause optimism ([#2508](https://github.com/synapsecns/sanguine/issues/2508)) ([4dda636](https://github.com/synapsecns/sanguine/commit/4dda636f21c70bb4f58e94616afc23b9ab9bdb12))
+
+
+
+
+
 ## [0.19.1](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.19.0...@synapsecns/synapse-interface@0.19.1) (2024-04-22)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface